requirements.txt
# Core Dependencies
python>=3.9
numpy>=1.24.0
pandas>=2.0.0

# Web Scraping & Crawling
scrapy>=2.11.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0
playwright>=1.40.0
lxml>=4.9.0

# NLP & ML
transformers>=4.35.0
torch>=2.1.0
sentencepiece>=0.1.99
hazm>=0.7.0
parsbert>=0.1.0
scikit-learn>=1.3.0

# LLM Support
openai>=1.3.0
langchain>=0.1.0
langchain-community>=0.0.10

# Data Processing
jsonschema>=4.19.0
pydantic>=2.5.0
python-dateutil>=2.8.0
pyyaml>=6.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
redis>=5.0.0

# API & Web
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2
aiohttp>=3.9.0
httpx>=0.25.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

